claude_generator/requirements.txt
aiofiles>=23.0.0
anthropic>=0.18.0
python-dotenv>=1.0.0
pyyaml>=6.0
httpx>=0.24.0
keyring>=24.0.0
cryptography>=41.0.0
psutil>=5.9.0
